% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utility_functions.R
\name{ascii_to_nor}
\alias{ascii_to_nor}
\title{Convert ASCII letters to Norwegian letters in psv files}
\usage{
ascii_to_nor(x)
}
\arguments{
\item{x}{vector}
}
\description{
Convert ASCII letters to Norwegian letters in psv files
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
\keyword{internal}
