% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBiotic_functions.R
\name{convertColumnTypes}
\alias{convertColumnTypes}
\title{Converts column types in a data frame to (hopefully) correct types}
\usage{
convertColumnTypes(df)
}
\arguments{
\item{k}{a data.table}
}
\value{
Returns a data.table with corrected column types. Also corrects misinterpreted Norwegian letters and dates.
}
\description{
Converts column types in a data frame to (hopefully) correct types
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
