% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBiotic_functions.R
\name{correctNorwegianLetters}
\alias{correctNorwegianLetters}
\title{Replace misinterpreted Norwegian letters by correct ones}
\usage{
correctNorwegianLetters(x)
}
\arguments{
\item{x}{character vector}
}
\value{
Returns a character vector with corrected letters
}
\description{
Replaces various misinterpretations of å, æ, and ø by the correct letters
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
