% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ers_example_data}
\alias{ers_example_data}
\title{Example ERS data}
\format{
A tibble
}
\source{
Norwegian Directorate of Fisheries (\url{https://www.fiskeridir.no/})
}
\usage{
ers_example_data
}
\description{
Example ERS data for beaked redfish. Use the \code{\link{extractERS}} function to produce similar data.
}
\keyword{datasets}
