% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractERS.R
\name{extractERS}
\alias{extractERS}
\title{Extract ERS information from detailed Norwegian fisheries logbooks}
\usage{
extractERS(
  path,
  species = NULL,
  translate_header = TRUE,
  method = "start",
  language = "norwegian",
  print.filename = FALSE,
  ncores = 1
)
}
\arguments{
\item{path}{Character argument defining the path to the folder where data are located. This can either be the entire folder in which case all the .zip files will be read or a single .zip file.}

\item{species}{Character argument defining the species. Species names can be given in Norwegian, English or Latin depending on the \code{language} argument. Species names are sampled from \code{\link{FDIRcodes}$speciesCodes}. \code{NULL} returns all species in the ERS data.}

\item{translate_header}{Logical indicating whether to translate the original column names in data selecting useful columns. If \code{FALSE}, the original columns are returned without modifications. If \code{TRUE}, columns are modified. See Return for details.}

\item{method}{Character argument specifying the method for position, time and depth extraction. Alternatives: \code{"start"} takes the reported start values, \code{"end"} extracts the reported end values and \code{"average"} takes an average of start and end values.}

\item{language}{Character argument in lower case specifying the language of species names in the returned data table. Alternatives: \code{"norwegian"}, \code{"english"}, or \code{"latin"}, The species names are acquired from \code{\link{FDIRcodes}$speciesCodes}.}

\item{print.filename}{Logical indicating whether the file name should be printed to console while processing. Useful for debugging.}

\item{ncores}{Integer giving the number of cores to be used when combining the files. Speeds up the calculus.}
}
\value{
If \code{translate_header = TRUE}, a data.table containing following columns:
\itemize{
\item \code{"year"} year of the fishing event.
\item \code{"month"} month of the fishing event.
\item \code{"id"} identification number of the fishing event.
\item \code{"vesselName"} name of the vessel.
\item \code{"radioCallSign"} call signal of the vessel.
\item \code{"EEZ"} reported exclusive economic zone of the fishing event.
\item \code{"ICESArea"} reported FAO area of the fishing event.
\item \code{"FDIRMainArea"} reported main area (hovedomraade) of the fishing event.
\item \code{"date"} date and time of the fishing event acquired based on the \code{method} argument.
\item \code{"lon"} longitude coordinate for the fishing event in decimal degrees acquired based on the \code{method} argument.
\item \code{"lat"} latitude coordinate for the fishing event in decimal degrees acquired based on the \code{method} argument.
\item \code{"depth"} reported fishing depth in meters acquired based on the \code{method} argument.
\item \code{"gear"} gear name translated from \code{gearId} based on \code{\link{FDIRcodes}$gearCodes}.
\item \code{"gearGroup"} reported gear group according to the Norwegian Directorate of Fisheries.
\item \code{"gearCategory"} gear category translated from \code{gearId} based on \code{\link{FDIRcodes}$gearCodes}.
\item \code{"gearSpecification"} reported gear specification according to the Norwegian Directorate of Fisheries.
\item \code{"gearMeshSize"} reported gear mesh size.
\item \code{"gearAmount"} reported gear amount (number of hooks etc.).
\item \code{"gearMeshSize"} reported issues with the gear.
\item \code{"fishTime"} reported fishing time in minutes.
\item \code{"fishDistance"} reported fishing distance in meters.
\item \code{"targetSp"} target species based on the FAO definition: the species with highest weight in catch.
\item \code{"catchSp"} reported catch species based on the \code{species}, \code{language} and \code{subspecies} arguments. Translated from the species ID (FANGSTART_NS or FANGSTART_FAO) based on \code{\link{FDIRcodes}$speciesCodes}.
\item \code{"weight"} reported catch of the \code{catchSp} in kilograms.
}

Note that NA catch species (\code{Art - FDIR}) are filtered out when \code{translate_header = TRUE}. This can include some records of fishing with no catch.

If \code{translate_header = FALSE}, the data are returned without filtering or modifications with the Norwegian column names.
}
\description{
Extracts information from detailed Norwegian fisheries logbook files downloaded using the \code{\link{downloadERS}} function.
}
\details{
This function extracts information from the Norwegian fisheries logbooks provided in the \code{.zip} format on the \href{https://www.fiskeridir.no/statistikk-tall-og-analyse/data-og-statistikk-om-yrkesfiske/apne-data-elektronisk-rapportering-ers}{Norwegian Directorate of Fisheries (FDIR) webpage}. Use the \code{\link{downloadERS}} function to download these data on your computer. Note that at the time of writing, FDIR only provided ERS data for > 15 m vessels due to unclarities in the Norwegian Privacy Act. If you want to access ERS data from smaller vessels, use the legacy function \code{\link{extractLogbook}}, which reads the confidential ERS data on IMR servers.
}
\seealso{
Other ERS functions:
\code{\link[=downloadERS]{downloadERS()}},
\code{\link[=extractLogbook]{extractLogbook()}}
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
\concept{ERS functions}
