% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{FDIRcodes}
\alias{FDIRcodes}
\title{List of Norwegian Directorate of Fisheries logbook codes}
\format{
A list of two data frames
}
\source{
Norwegian Directorate of Fisheries (\url{https://www.fiskeridir.no})
}
\usage{
data(FDIRcodes)
}
\description{
A named list with two elements: \code{$speciesCodes} (FDIR species codes with FAO codes, Norwegian, English, and Latin names) and \code{$gearCodes} (FDIR gear codes with Norwegian/English categories). Use \code{\link{readFdirCodes}} to update from the FDIR Excel code list.
}
\keyword{datasets}
