% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{fishingAreasNor}
\alias{fishingAreasNor}
\title{Major fishing zones of Norway}
\format{
\code{\link[sp:SpatialPolygons]{SpatialPolygonsDataFrame}} in decimal degrees (+init=epsg:4326) containing major fishing zones defined by the Norwegian Directorate of Fisheries.
}
\source{
\href{https://kart.fiskeridir.no/stat}{Norwegian Directorate of Fisheries}
}
\usage{
fishingAreasNor
}
\description{
Major fishing zones of Norway
}
\concept{shapefiles}
\keyword{datasets}
\keyword{shapefiles}
