% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalPolygonizeR.R
\name{gdalPolygonizeR}
\alias{gdalPolygonizeR}
\title{Polygonize raster using a Python script and QGIS}
\usage{
gdalPolygonizeR(
  x,
  outshape = NULL,
  gdalformat = "ESRI Shapefile",
  pypath = NULL,
  readpoly = TRUE,
  quiet = TRUE,
  rewrite = TRUE
)
}
\arguments{
\item{x}{raster object to be polygonized}

\item{outshape}{The path where the converted polygon shapefile will reside. Default is \code{NULL}, which saves the shapefile to a temporary file}

\item{gdalformat}{...}

\item{pypath}{...}

\item{readpoly}{Logical indicating}

\item{quiet}{...}

\item{rewrite}{...}
}
\description{
...
}
\details{
Documentation and code tidying has not been written for this function yet.
}
\author{
\href{https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/}{John Baumgartner}. See also \href{https://gis.stackexchange.com/questions/166753/fastest-way-to-convert-big-raster-to-polyline-using-r-or-python}{StackOverflow}.
}
