% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{gearList}
\alias{gearList}
\title{NMD/IMR gear code list}
\format{
A data table
}
\source{
Institute of Marine Research (\url{https://imr.no})
}
\usage{
gearList
}
\description{
Use the \code{\link[BioticExplorerServer]{prepareGearList}} function to update the list.
}
\keyword{datasets}
