% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainAreaCentroid.R
\name{mainAreaCentroid}
\alias{mainAreaCentroid}
\title{Get centroid of a Norwegian main fishing area}
\usage{
mainAreaCentroid(name, return.sp = FALSE)
}
\arguments{
\item{name}{Integer vector giving the names of main areas}

\item{return.sp}{Logical indicating whether a SpatialPoints object or data frame should be returned.}
}
\value{
A SpatialPoints object if \code{return.sp = TRUE}. Otherwise, a data frame
}
\description{
Returns a centroid of a Norwegian Directorate of Fisheries main fishing area (hoverområde) in decimal degrees
}
\examples{
mainAreaCentroid(1:70)
}
\author{
Mikko Vihtakari
}
