% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{nea_ghl_original_strata}
\alias{nea_ghl_original_strata}
\title{North East Atlantic Greenland original strata areas}
\format{
A dataframe
}
\source{
Institute of Marine Research (\url{https://imr.no})
}
\usage{
data(nea_ghl_original_strata)
}
\description{
North East Atlantic Greenland original strata areas
}
\keyword{datasets}
