% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{nea_ghl_survey_stations}
\alias{nea_ghl_survey_stations}
\title{North East Atlantic Greenland halibut survey stations}
\format{
A dataframe
}
\source{
Institute of Marine Research (\url{https://www.imr.no})
}
\usage{
data(nea_ghl_survey_stations)
}
\description{
North East Atlantic Greenland halibut survey stations
}
\keyword{datasets}
