% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utility_functions.R
\name{nor_to_ascii}
\alias{nor_to_ascii}
\title{Convert Norwegian letters to ASCII equivalents}
\usage{
nor_to_ascii(x)
}
\arguments{
\item{x}{vector}
}
\description{
Convert Norwegian letters to ASCII letters
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
\keyword{internal}
