% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOnFishingArea.R
\name{pointOnFishingArea}
\alias{pointOnFishingArea}
\title{Detect statistical fishing area from decimal degree coordinates}
\usage{
pointOnFishingArea(lon, lat, type = "fdir")
}
\arguments{
\item{lon}{Numeric vector of longitude coordinates as decimal degrees. Must be the same length than \code{lat}}

\item{lat}{Numeric vector of latitude coordinates as decimal degrees. Must be the same length than \code{lon}}

\item{type}{A character argument specifying the fisheries area type. Alternatives: \code{"fdir"} for the Norwegian Directorate of Fisheries main fishing areas, \code{"ices"} for FAO/ICES fisheries areas.}
}
\value{
Returns a an integer vector with names of areas. Same length than \code{lon} and \code{lat}. \code{NA} is returned, if a point is outside the areas.
}
\description{
Detects Norwegian Directorate of Fisheries main fishing areas (hoverområde) and FAO/ICES fisheries areas from decimal degree coordinates.
}
\details{
The function is designed to work with Norwegian data. All FAO fisheries areas are not included.
}
\examples{
dt <- data.frame(lon = seq(-50, 50, 10), 
                 lat = seq(50, 80, length.out = length(lon)))
dt$main_area <- pointOnFishingArea(dt$lon, dt$lat)
dt$ices_area <- pointOnFishingArea(dt$lon, dt$lat, type = "ices")
dt[is.na(dt)] <- "NA"

\dontrun{
# Norwegian fishing areas

labels <- sp::SpatialPointsDataFrame(rgeos::gCentroid(fishingAreasNor, byid=TRUE), 
                                     data = fishingAreasNor@data)
labels <- ggspatial::df_spatial(labels)

ggOceanMaps::basemap(limits = raster::extent(fishingAreasNor)[1:4]) +
  ggspatial::annotation_spatial(fishingAreasNor, fill = NA, color = "blue") +
  ggspatial::geom_spatial_text(data = dt, 
                               aes(x = lon, y = lat, label = main_area), 
                               size = FS(10), fontface = 2, color = "red") + 
  ggspatial::geom_spatial_text(data = labels, aes(x = x, y = y, label = FID),
                               size = FS(8), fontface = 2, color = "blue") 

# ICES areas

ggOceanMaps::basemap(data = dt) +
  ggspatial::annotation_spatial(icesAreas, fill = NA, color = "blue") +
  ggspatial::geom_spatial_text(data = dt, 
                               aes(x = lon, y = lat, label = ices_area), 
                               size = FS(10), fontface = 2, color = "red")

}
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
