% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOnMainArea.R
\name{pointOnMainArea}
\alias{pointOnMainArea}
\title{Detect Norwegian main fishing area from decimal degree coordinates}
\usage{
pointOnMainArea(lon, lat)
}
\arguments{
\item{lon}{Numeric vector of longitude coordinates as decimal degrees. Must be the same length than \code{lat}}

\item{lat}{Numeric vector of latitude coordinates as decimal degrees. Must be the same length than \code{lon}}
}
\value{
Returns a an integer vector with names of main areas. Same length than \code{lon} and \code{lat}. \code{NA} is returned, if a point is outside the areas.
}
\description{
Detects Norwegian Directorate of Fisheries main fishing area (hoverområde) from decimal degree coordinates
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
