% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes_reexports.R
\name{prepareCruiseSeriesList}
\alias{prepareCruiseSeriesList}
\title{Prepare cruise series list}
\usage{
prepareCruiseSeriesList()
}
\description{
Wrapper for \code{\link[BioticExplorerServer]{prepareCruiseSeriesList}}.
Uses the updated IMR reference API. See BioticExplorerServer for full documentation.
}
