% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes_reexports.R
\name{prepareGearList}
\alias{prepareGearList}
\title{Prepare gear list}
\usage{
prepareGearList()
}
\description{
Wrapper for \code{\link[BioticExplorerServer]{prepareGearList}}.
Uses the updated IMR reference API. See BioticExplorerServer for full documentation.
}
