% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utility_functions.R
\name{pretty_log}
\alias{pretty_log}
\title{Round to pretty log breaks}
\usage{
pretty_log(x, f = round)
}
\arguments{
\item{x}{numeric vector to round}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
Rounds values to human-readable breaks suitable for logarithmic scales.
}
\seealso{
\code{\link{round_any}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
