% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBiotic_functions.R
\name{processBioticFile}
\alias{processBioticFile}
\title{Read and process a NMD Biotic xml file for further use in the BioticExplorer}
\usage{
processBioticFile(
  file,
  lengthUnit = "cm",
  weightUnit = "g",
  removeEmpty = TRUE,
  coreDataOnly = FALSE,
  returnOriginal = TRUE,
  dataTable = TRUE,
  convertColumns = FALSE,
  missionidPrefix = NULL
)
}
\arguments{
\item{file}{character string specifying the file path to the xml file. Accepts only one file at the time.}

\item{lengthUnit}{character string specifying the unit for length output. Alternatives: "mm", "cm" or "m".}

\item{weightUnit}{character string specifying the unit for weight output. Alternatives: "g" or "kg".}

\item{removeEmpty}{logical indicating whether empty columns should be removed from the output. This option also influences "coreData" columns.}

\item{coreDataOnly}{logical indicating whether only important core columns should be picked from data. See \code{\link{coreDataList}} for list of core columns for each data type.}

\item{returnOriginal}{logical indicating whether the original data (\code{$mission} through \code{$agedetermination}) should be returned together with combined data.}

\item{dataTable}{logical indicating whether the output should be returned as \link[data.table]{data.table}s instead of \link{data.frame}s. Setting this to \code{TRUE} speeds up further calculations using the data (but requires the \link[data.table]{data.table} syntax).}

\item{convertColumns}{logical indicating whether the column types should be converted. See \code{link{convertColumnTypes}}. Setting this to \code{FALSE} considerably speeds up the function, but leads to problems with non-unicode characters.}

\item{missionidPrefix}{A prefix for the \code{missionid} identifier, which separates cruises. Used in \code{\link{processBioticFiles}} function when several xml files are put together. \code{NULL} (default) omits the prefix. Not needed in \code{processBioticFile} function.}
}
\value{
Returns a list of Biotic data with \code{$mission}, \code{$fishstation}, \code{$catchsample}, \code{$individual} and \code{$agedetermination} data frames. The \code{$stnall} and \code{$indall} data frames are merged from \code{$fishstation} and \code{$catchsample} (former) and  \code{$fishstation}, \code{$catchsample}, \code{$individual} and \code{$agedetermination} (latter).
}
\description{
A wrapper for \code{\link[RstoxData]{readXmlFile}} to enable further use in the BioticExplorer
}
\seealso{
Other Biotic functions:
\code{\link[=print.bioticProcData]{print.bioticProcData()}},
\code{\link[=processBioticFiles]{processBioticFiles()}}
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
\concept{Biotic functions}
