% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utility_functions.R
\name{quiet}
\alias{quiet}
\title{Return function output quietly}
\usage{
quiet(x)
}
\arguments{
\item{x}{function}
}
\value{
The return value of \code{x}, invisibly.
}
\description{
Returns function output without printed \code{\link{cat}} messages
}
\author{
Hadley Wickham
}
\keyword{internal}
