% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.pos.R
\name{read.pos}
\alias{read.pos}
\title{Read and compile IMR vessel position files from a folder}
\usage{
read.pos(path, log.file = TRUE)
}
\arguments{
\item{path}{Character string defining the path to the code{/CRUISE_LOG/TRACK/} folder}

\item{log.file}{Logical indicating whether the file from \code{path} has \code{.log} extension. If false, \code{.csv} is assumed.}
}
\value{
A data.table containing all position information from the code{/CRUISE_LOG/TRACK/} folder
}
\description{
Compiles the standard IMR vessel position files into a data.table for further analyses
}
\author{
Mikko Vihtakari
}
