% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSluttseddelXLS.R
\name{readSluttseddelXLS}
\alias{readSluttseddelXLS}
\title{Read landing data from annually compiled Excel files}
\usage{
readSluttseddelXLS(
  species,
  dataDir,
  years = 1977:2022,
  dropMissingMainArea = TRUE
)
}
\arguments{
\item{species}{Character string defining the species using the FDIR Norwegian species names. \code{NULL} returns all species.}

\item{dataDir}{Character vector defining the path to the folder where Excel files are located (typically "sluttseddel_xls_ferdige_År")}

\item{years}{Integer vector defining the years to read. Files in later years tend to be a mess and this option avoids crashes. Use \code{NULL} to try opening all files.}

\item{dropMissingMainArea}{Logical indicating whether catches with missing main area should be dropped.}
}
\value{
A data.frame with columns \code{year}, \code{month}, \code{main_area}, \code{sub_area}, \code{ices_area}, \code{gear_id}, \code{gear_category}, \code{gear}, and \code{weight} (and \code{species} when \code{species = NULL}).
}
\description{
Reads and compiles landing (sales note / sluttseddel) data from the annually compiled Excel files stored on the IMR server. Requires access to the IMR intranet and the Excel files to be in place.
}
\seealso{
Other Landings functions:
\code{\link[=downloadLandings]{downloadLandings()}}
}
\concept{Landings functions}
