% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utility_functions.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric vector to round}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
Rounds values to the nearest multiple of a given accuracy.
}
\author{
Hadley Wickham
}
\keyword{internal}
