% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{salesnote_example_data}
\alias{salesnote_example_data}
\title{Example sales note data downloaded through the API}
\format{
A tibble
}
\source{
Norwegian Directorate of Fisheries (\url{https://www.fiskeridir.no/})
}
\usage{
salesnote_example_data
}
\description{
Example sales note data for beaked redfish. Use the \code{\link{downloadLandings}} function to produce similar data.
}
\keyword{datasets}
