% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{salesnote_xls_data}
\alias{salesnote_xls_data}
\title{Example historic sales note data extracted from Excel sheets on IMR server}
\format{
A tibble
}
\source{
Institute of Marine Research (\url{https://imr.no}) and Norwegian Directorate of Fisheries (\url{https://www.fiskeridir.no/})
}
\usage{
salesnote_xls_data
}
\description{
Example sales note data for beaked redfish. Use the \code{\link{readSluttseddelXLS}} function to produce similar data.
}
\keyword{datasets}
