% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utility_functions.R
\name{select.element}
\alias{select.element}
\title{Select an element of each vector from a list}
\usage{
select.element(x, y)
}
\arguments{
\item{x}{list}

\item{y}{number of element. Must be integer}
}
\value{
A vector of the y'th element from each vector in \code{x}.
}
\description{
Selects y'th element of each vector from a list
}
\keyword{internal}
