% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataPolygon.R
\name{strataPolygon}
\alias{strataPolygon}
\title{Create polygon shapefiles for depth strata required by Rstox}
\usage{
strataPolygon(bathy, depths, boundary, geostrata = NULL,
  drop.crumbs = NULL, remove.holes = NULL, strata.names = NULL,
  validate.polygons = TRUE, use.python = FALSE)
}
\arguments{
\item{bathy}{String giving the path to the bathymetry NetCDF file.}

\item{depths}{Numeric vector giving the cut points for depth strata (see \code{\link[base]{cut}}. Data outside the cut range will be dropped. Use limits of length two exceeding the depths of the region to avoid depth categorization (\code{c(0, 1000)} for instance).}

\item{boundary}{Numeric vector of length 4 giving the boundaries for the overall region. See \code{\link[raster]{extent}}. Should be given as decimal degrees. The first element defines the minimum longitude, the second element the maximum longitude, the third element the minimum latitude and the fourth element the maximum latitude of the bounding box.}

\item{geostrata}{A data frame defining the minimum and maximum longitude and latitude for geographically bounded strata. The data frame columns must be ordered as \code{lon.min, lon.max, lat.min, lat.max}. Column names do not matter. Each row in the data frame will be interpreted as separate geographically bounded strata. Use \code{NULL} to ignore geostrata.}

\item{drop.crumbs}{Single numeric value specifying a threshold (area in km2) for disconnected polygons which should be removed from the strata. Set to \code{NULL} to bypass the removal. Uses the \link[smoothr]{drop_crumbs} function.}

\item{remove.holes}{Single numeric value specifying a threshold (area in km2) for holes which should be removed from the strata. Set to \code{NULL} to bypass the removal. Uses the \link[smoothr]{fill_holes} function. 
A single logical argument or a logical vector as long as the number of rows in \code{geostrata} specifying whether holes in the strata polygons should be removed.}

\item{strata.names}{A character vector spcifying the names of \code{geostrata}. Not implemented.}

\item{validate.polygons}{A logical indicating whether the function should return valid geometries. This option might considerably change the output, but makes it compatible with GIS software.}

\item{use.python}{Logical indicating whether the function should use gdal python script (\code{TRUE}; \code{gdal_polygonize.py}) or \code{\link[stars]{st_as_stars}} (\code{FALSE}) for polygonization of strata. The python script has a superior computing time, but requires QGIS 2.18 installed on the computer (earlier or later versions won't do).}
}
\value{
\link[sp]{SpatialPolygonsDataFrame} containing the estimated strata and information for them including areas. The strata are returned as decimal degrees (WGS84).
}
\description{
A helper function to define strata for stock assesment from GEBCO and ETOPO bathymetry grids.
}
\details{
Uses \href{https://www.gebco.net/data_and_products/gridded_bathymetry_data/}{GEBCO} or \href{https://www.ngdc.noaa.gov/mgg/global/}{ETOPO1} bathymetry grids to define the depth strata. Download the desired grid from the links. The bathymetry grids must be in NetCDF format.
}
\references{
GEBCO Compilation Group (2019) GEBCO 2019 15-arcsecond grid (doi:10.5285/836f016a-33be-6ddc-e053-6c86abc0788e). URL: \url{https://www.gebco.net/data_and_products/gridded_bathymetry_data/gebco_2019/gebco_2019_info.html}.

ETOPO1 1 Arc-Minute Global Relief Model. \url{https://doi.org/10.7289/V5C8276M}.
}
\seealso{
\code{\link{gdalPolygonizeR}} for the polygonization algorithm. 
\code{\link{strataArea}} for calculating strata areas without polygonization.
}
\author{
Mikko Vihtakari
}
