% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStrataPolygonToWKT.R
\name{strataPolygonToWKT}
\alias{strataPolygonToWKT}
\title{Convert strata spatial polygons to the WKT format required by Rstox}
\usage{
strataPolygonToWKT(x)
}
\arguments{
\item{x}{A SpatialPolygonsDataFrame containing the strata polygons.}
}
\value{
...
}
\description{
Converts strata polygons from \code{\link{strataPolygon}} to the WKT format required by Rstox
}
\author{
Mikko Vihtakari
}
