% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{taxaList}
\alias{taxaList}
\title{NMD/IMR taxa code list}
\format{
A data table
}
\source{
Institute of Marine Research (\url{https://imr.no})
}
\usage{
taxaList
}
\description{
Use the \code{\link[BioticExplorerServer]{prepareTaxaList}} function to update the list.
}
\keyword{datasets}
