% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCoord.R
\name{transformCoord}
\alias{transformCoord}
\title{Transform spatial coordinates to another projection}
\usage{
transformCoord(
  x = NULL,
  lon = NULL,
  lat = NULL,
  new.names = c("lon.utm", "lat.utm"),
  proj.og = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  proj.out = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  verbose = FALSE,
  bind = FALSE,
  na = "ignore"
)
}
\arguments{
\item{x}{data frame to be transformed. Can be omitted if numeric vectors are assigned to \code{lon} and \code{lat}.}

\item{lon, lat}{either a name of the longitude and latitude columns in \code{x} or a numeric vector containing longitude and latitude coordinates. Use \code{NULL} to guess the longitude and/or latitude columns in \code{x}.}

\item{new.names}{a vector of length 2 specifying the names of transformed longitude and latitude columns, respectively. \code{NULL} returns column names from \code{x}}

\item{proj.og}{original \code{\link[sp]{proj4string}} projection. If \code{NULL}, the projection is taken from \code{x}. \code{x} must be a \link[sp]{Spatial} object in that case.}

\item{proj.out}{the \code{\link[sp]{proj4string}} projection the coordinates should be transformed to. Defaults to \code{proj.og} and no transformation will be made.}

\item{verbose}{if \code{TRUE}, the function prints information about the changed projection. Switch to \code{FALSE} to make the function silent.}

\item{bind}{logical. Should only transformed coordinates be returned (\code{FALSE}, default) or should x be returned with transformed coordinates (\code{TRUE})?}

\item{na}{character specifying the NA action for missing coordinates. The "ignore" option ignores the coordinates and returns NAs to transformed coordinates. The "remove" option removes missing values from \code{x} returning a message while doing it. Any other character argument will tricker \code{na.fail} stopping the function in case of missing coordinates.}
}
\value{
Returns a data frame with transformed spatial coordinates
}
\description{
The function transforms spatial coordinates from original projection (decimal degrees assumed) to another projection.
}
\details{
If \code{x} is specified, the function guesses longitude and latitude columns from \code{x} by default.
}
\author{
Mikko Vihtakari
}
