% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBiotic_functions.R
\name{tryCatchWE}
\alias{tryCatchWE}
\title{tryCatch both warnings (with value) and errors}
\usage{
tryCatchWE(expr)
}
\arguments{
\item{expr}{an \R expression to evaluate}
}
\value{
List of logicals indicating whether the \code{expr} produces a warning or error.
}
\description{
tryCatch both warnings (with value) and errors
}
\author{
Martin Maechler; Copyright (C) 2010-2012 The R Core Team, Mikko Vihtakari (Institute of Marine Research)
}
\keyword{internal}
