% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBiotic_functions.R
\name{coreDataList}
\alias{coreDataList}
\title{List of core data columns by data type in NMD Biotic data}
\usage{
coreDataList(type)
}
\arguments{
\item{type}{character string specifying the data type. Alternatives: "mission", "fishstation", "individual", "catchsample", or "agedetermination".}
}
\value{
Returns a character vector of core data types for a given data type.
}
\description{
List of core data types used in the \code{\link{processBioticFile}} function
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
\keyword{internal}
