% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{cruiseSeriesList}
\alias{cruiseSeriesList}
\title{NMD/IMR cruise series list}
\format{
A data table
}
\source{
Institute of Marine Research (\url{https://imr.no})
}
\usage{
cruiseSeriesList
}
\description{
Use the \code{\link{prepareCruiseSeriesList}} function to update the list. Need to be updated every time data from the database are downloaded.
}
\keyword{datasets}
