% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadERS.R
\name{downloadERS}
\alias{downloadERS}
\title{Download Electronic Reporting System (ERS) catch data into a folder from the \href{https://www.fiskeridir.no/statistikk-tall-og-analyse/data-og-statistikk-om-yrkesfiske/apne-data-elektronisk-rapportering-ers}{Norwegian Directorate of Fisheries (FDIR) webpage}}
\usage{
downloadERS(
  dest,
  years = 2011:as.integer(format(Sys.Date(), "\%Y")),
  overwrite = "yes",
  fdir_url = "https://register.fiskeridir.no/vms-ers/ERS/"
)
}
\arguments{
\item{dest}{File path as character where the data should be downloaded to.}

\item{years}{An integer vector of years to download. The default downloads all years.}

\item{overwrite}{Either \code{"yes"}, \code{"no"} or \code{"force"}. The "yes" option overwrites annual .zip files only if the were modified on the webpage compared to existing files, "no" returns an error if there are existing .zip files in \code{dest} and "force" downloads all .zip files from the website again overwriting all existing files.}

\item{fdir_url}{Character defining the URL where data should be downloaded from. You'll only need to modify this if the source URL has changed. In that case, send an email to the maintainer such that the address can be updated.}
}
\value{
Returns nothing. Downloads data into folder specified by the \code{dest} argument.
}
\description{
Download Electronic Reporting System (ERS) catch data into a folder from the \href{https://www.fiskeridir.no/statistikk-tall-og-analyse/data-og-statistikk-om-yrkesfiske/apne-data-elektronisk-rapportering-ers}{Norwegian Directorate of Fisheries (FDIR) webpage}
}
\details{
Please note that this function downloads large amounts of data and places then as zip files in the folder defined by \code{dest}.
}
\seealso{
Other ERS functions: 
\code{\link{extractERS}()},
\code{\link{extractLogbook}()}
}
\author{
Mikko Vihtakari
}
\concept{ERS functions}
