% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadLandings.R
\name{downloadLandings}
\alias{downloadLandings}
\title{Download landings data for a species from the IMR database}
\usage{
downloadLandings(species, years = NULL)
}
\arguments{
\item{species}{any species identification name in \code{FDIRcodes$speciesCodes} as character. Only one species at the time allowed.}

\item{years}{an integer vector of years to download. If \code{NULL} (default), all years are downloaded. Please note that this option can take very long time and lead to huge datasets.}
}
\description{
The function downloads landings ("sluttseddel") data from IMR database. Requires access to the intranet.
}
\examples{
\dontrun{
downloadLandings("brugde") # Basking shark, all years
downloadLandings("kveite", years = 2000:2001) # halibut, 2000-01
}
}
\author{
Mikko Vihtakari
}
