% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadLandings.R
\name{downloadLandings}
\alias{downloadLandings}
\title{Download landings data for a species from the IMR database}
\usage{
downloadLandings(species, years = c(1900:2020))
}
\arguments{
\item{species}{any species identification name in \code{FDIRcodes$speciesCodes} as character. Only one species at the time allowed.}

\item{years}{an integer vector of years to download. Defaults to all years. Please note that this option can take very long time and lead to huge datasets.}
}
\description{
THe function downloads landings ("sluttseddel") data from IMR database. Requires access to the intranet.
}
\author{
Mikko Vihtakari
}
