% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLogbook.R
\name{extractLogbook}
\alias{extractLogbook}
\title{Extract information from detailed Norwegian fisheries logbooks}
\usage{
extractLogbook(path, species, method = "start", remove.sensitive = TRUE)
}
\arguments{
\item{path}{Character argument defining the path to the folder where data are located}

\item{species}{Character argument defining the species. Species names are given in Norwegian (see Håndbok for prøvetaking). Some of the available names are peculiar, but most commercially valuable species should work.}

\item{method}{Character argument specifying the method for position, time and depth extraction. Alternatives: \code{"start"} takes the reported start values, \code{"end"} extracts the reported end values and \code{"average"} takes an average of start and end values.}

\item{remove.sensitive}{Logical indicating whether the function should remove sensitive data fields.}
}
\value{
A data.table containing the desired information.
}
\description{
Extracts information from detailed Norwegian fisheries logbook files
}
\details{
This function extracts information from the Norwegian fisheries logbooks provided in the \code{.psv} format on the IMR server. Note that these data are confidential and protected by the Norwegian Privacy Act (personvernloven). Keep the data on the server or in another safe place (not on the cloud). Do not send the input data around on email. The \code{remove.sensitive} argument can be used to disconnect the output from the confidential input data. This output can be distributed to people who have not signed the confidentiality statement (students, ICES colleagues, etc.) also using email. The data are property of the Norwegian Directorate of Fisheries and you should consult them if you plan to publish any parts of the dataset. 

Note that the structure of logbook files is not confidential and therefore this function can be distributed freely. The data required by the function cannot (do not send those data on email).

The mess with species names may be solved in the future. For now, you just have to know what a species you are interested in is called in these data.
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
