% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{FDIRcodes}
\alias{FDIRcodes}
\title{List of Norwegian Directorate of Fisheries lobbook codes}
\format{
A list of data tables
}
\source{
Norwegian Directorate of Fisheries (\url{https://www.fiskeridir.no/English})
}
\usage{
data(FDIRcodes)
}
\description{
List of Norwegian Directorate of Fisheries lobbook codes
}
\keyword{datasets}
