% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{gearList}
\alias{gearList}
\title{Gear list for Norwegian Fisheries Logbook data}
\format{
A data table
}
\source{
Institute of Marine Research (\url{https://imr.no})
}
\usage{
data(gearList)
}
\description{
Gear list for Norwegian Fisheries Logbook data
}
\keyword{datasets}
