% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCruiseSeriesList.R
\name{prepareCruiseSeriesList}
\alias{prepareCruiseSeriesList}
\title{Prepare cruise series list}
\usage{
prepareCruiseSeriesList()
}
\description{
Prepares cruise series list using the IMR database API. Stored as data object in the package. Refreshes any new cruise codes. The function does not use any arguments.
}
\author{
Ibrahim Umar, Mikko Vihtakari
}
