% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes_reexports.R
\name{prepareGearList}
\alias{prepareGearList}
\title{Prepare gear list}
\usage{
prepareGearList()
}
\value{
A data.table of gear codes and categories from the IMR reference API.
}
\description{
Wrapper for \code{\link[BioticExplorerServer]{prepareGearList}}.
Uses the updated IMR reference API. See BioticExplorerServer for full documentation.
}
\seealso{
Other Biotic functions:
\code{\link[=prepareCruiseSeriesList]{prepareCruiseSeriesList()}},
\code{\link[=prepareTaxaList]{prepareTaxaList()}},
\code{\link[=print.bioticProcData]{print.bioticProcData()}},
\code{\link[=processBioticFile]{processBioticFile()}},
\code{\link[=processBioticFiles]{processBioticFiles()}}
}
\concept{Biotic functions}
