% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGearList.R
\name{prepareGearList}
\alias{prepareGearList}
\title{Prepare NMD/IMR gear list}
\usage{
prepareGearList()
}
\description{
Prepares gear list using the IMR database API. Stored as data object in the package. Refreshes any new gear codes. The function does not use any arguments.
}
\author{
Mikko Vihtakari, Ibrahim Umar
}
