% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes_reexports.R
\name{prepareTaxaList}
\alias{prepareTaxaList}
\title{Prepare taxa list}
\usage{
prepareTaxaList(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Print status messages during download. Defaults to \code{FALSE}.}
}
\description{
Wrapper for \code{\link[BioticExplorerServer]{prepareTaxaList}}.
Uses the updated IMR reference API. See BioticExplorerServer for full documentation.
}
