% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBiotic_functions.R
\name{print.bioticProcData}
\alias{print.bioticProcData}
\title{Print processed NMD Biotic data (\code{bioticProcData}) objects}
\usage{
\method{print}{bioticProcData}(x, ...)
}
\arguments{
\item{x}{\code{bioticProcData} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=processBioticFile]{bioticProcData}} objects
}
\seealso{
\code{\link{processBioticFile}} \code{\link{processBioticFiles}}
}
\author{
Mikko Vihtakari
}
