% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSluttseddelXLS.R
\name{readSluttseddelXLS}
\alias{readSluttseddelXLS}
\title{Read landing data from annually compiled Excel files}
\usage{
readSluttseddelXLS(species, dataDir)
}
\arguments{
\item{species}{Character string defining the species using the FDIR Norwegian species names}

\item{dataDir}{Character vector defining the path to the folder where Excel files are located (typically "sluttseddel_xls_ferdige_År")}
}
\description{
Read landing data from annually compiled Excel files
}
