% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RstoxUtils.R
\docType{package}
\name{RstoxUtils-package}
\alias{RstoxUtils}
\alias{RstoxUtils-package}
\title{Utility functions for the Stox Project}
\description{
This package contains utility functions for the Institute of Marine Research's (IMR) 
Stox Project. The package has two purposes: 1) To function as a showcase and
developmental platform for functions that may be included in the future releases of
the Stox Project. 2) To provide a collection of functions needed in the internal workflow 
of the Deep-sea species group at IMR. See the 
\href{https://github.com/MikkoVihtakari/RstoxUtils}{README} for further details
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MikkoVihtakari/RstoxUtils}
  \item Report bugs at \url{https://github.com/MikkoVihtakari/RstoxUtils/issues}
}

}
\author{
\strong{Maintainer}: Mikko Vihtakari \email{mikko.vihtakari@hi.no}

}
